Semester combobox component initial setup
